Add auth controller student and teacher

+++ src/main/kotlin/unq/pds/api/controller/StudentController.kt
@@ -1,5 +1,6 @@
 package unq.pds.api.controller
 
+import io.jsonwebtoken.Jwts
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.media.ArraySchema
 import io.swagger.v3.oas.annotations.media.Content
@@ -13,7 +14,6 @@ import org.springframework.web.bind.annotation.*
 import unq.pds.api.dtos.MessageDTO
 import unq.pds.api.dtos.StudentCreateRequestDTO
 import unq.pds.model.Student
-import unq.pds.model.exceptions.AlreadyRegisteredException
 import unq.pds.services.StudentService
 import javax.validation.Valid
 import javax.validation.constraints.NotBlank
@@ -54,8 +54,19 @@ class StudentController {
                     )]
                 )]
         )
-        fun createStudent(@RequestBody @Valid student: StudentCreateRequestDTO): ResponseEntity<Student> {
-            return ResponseEntity(studentService.save(student), HttpStatus.OK)
+        fun createStudent(
+            @CookieValue("jwt") jwt: String?,
+            @RequestBody @Valid student: StudentCreateRequestDTO
+        ): ResponseEntity<Any> {
+            if (jwt.isNullOrBlank()) {
+                return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+            }
+            val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+            return if (body["role"] == "STUDENT") ResponseEntity(
+                MessageDTO("You do not have permissions to access this resource"),
+                HttpStatus.UNAUTHORIZED
+            )
+            else ResponseEntity(studentService.save(student), HttpStatus.OK)
         }
 
         @GetMapping
@@ -98,7 +109,10 @@ class StudentController {
                     )]
                 )]
         )
-        fun getStudent(@NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+        fun getStudent(@CookieValue("jwt") jwt: String?, @NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+            if (jwt.isNullOrBlank()) {
+                return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+            }
             return ResponseEntity(studentService.findById(id), HttpStatus.OK)
         }
 
@@ -142,8 +156,16 @@ class StudentController {
                     )]
                 )]
         )
-        fun updateStudent(@RequestBody student: Student): ResponseEntity<Any> {
-            return ResponseEntity(studentService.update(student), HttpStatus.OK)
+        fun updateStudent(@CookieValue("jwt") jwt: String?, @RequestBody student: Student): ResponseEntity<Any> {
+            if (jwt.isNullOrBlank()) {
+                return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+            }
+            val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+            return if (body["role"] == "STUDENT") ResponseEntity(
+                MessageDTO("You do not have permissions to access this resource"),
+                HttpStatus.UNAUTHORIZED
+            )
+            else ResponseEntity(studentService.update(student), HttpStatus.OK)
         }
 
         @DeleteMapping
@@ -188,7 +210,15 @@ class StudentController {
                     )]
                 )]
         )
-        fun deleteStudent(@NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+        fun deleteStudent(@CookieValue("jwt") jwt: String?, @NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+            if (jwt.isNullOrBlank()) {
+                return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+            }
+            val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+            if (body["role"] == "STUDENT") return ResponseEntity(
+                MessageDTO("You do not have permissions to access this resource"),
+                HttpStatus.UNAUTHORIZED
+            )
             studentService.deleteById(id)
             return ResponseEntity(MessageDTO("Student has been deleted successfully"), HttpStatus.OK)
         }
@@ -233,8 +263,20 @@ class StudentController {
                     )]
                 )]
         )
-        fun addProject(@NotBlank @PathVariable studentId: Long, @NotBlank @PathVariable projectId: Long): ResponseEntity<Any> {
-            return ResponseEntity(studentService.addProject(studentId, projectId), HttpStatus.OK)
+        fun addProject(
+            @CookieValue("jwt") jwt: String?,
+            @NotBlank @PathVariable studentId: Long,
+            @NotBlank @PathVariable projectId: Long
+        ): ResponseEntity<Any> {
+            if (jwt.isNullOrBlank()) {
+                return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+            }
+            val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+            return if (body["role"] == "STUDENT") ResponseEntity(
+                MessageDTO("You do not have permissions to access this resource"),
+                HttpStatus.UNAUTHORIZED
+            )
+            else ResponseEntity(studentService.addProject(studentId, projectId), HttpStatus.OK)
         }
 
         @GetMapping("/getAll")
@@ -255,7 +297,10 @@ class StudentController {
                     ]
                 )]
         )
-        fun getAll(): ResponseEntity<List<Student>> {
+        fun getAll(@CookieValue("jwt") jwt: String?): ResponseEntity<Any> {
+            if (jwt.isNullOrBlank()) {
+                return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+            }
             return ResponseEntity(studentService.readAll(), HttpStatus.OK)
         }
     }
