+++ src/main/kotlin/unq/pds/api/controller/TeacherController.kt
@@ -1,5 +1,6 @@
 package unq.pds.api.controller
 
+import io.jsonwebtoken.Jwts
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.media.ArraySchema
 import io.swagger.v3.oas.annotations.media.Content
@@ -13,7 +14,6 @@ import org.springframework.web.bind.annotation.*
 import unq.pds.api.dtos.MessageDTO
 import unq.pds.api.dtos.TeacherCreateRequestDTO
 import unq.pds.model.Teacher
-import unq.pds.model.exceptions.AlreadyRegisteredException
 import unq.pds.services.TeacherService
 import javax.validation.Valid
 import javax.validation.constraints.NotBlank
@@ -54,8 +54,19 @@ class TeacherController {
                     )]
                 )]
         )
-        fun createTeacher(@RequestBody @Valid teacher: TeacherCreateRequestDTO): ResponseEntity<Teacher> {
-            return ResponseEntity(teacherService.save(teacher), HttpStatus.OK)
+        fun createTeacher(
+            @CookieValue("jwt") jwt: String?,
+            @RequestBody @Valid teacher: TeacherCreateRequestDTO
+        ): ResponseEntity<Any> {
+            if (jwt.isNullOrBlank()) {
+                return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+            }
+            val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+            return if (body["role"] == "STUDENT") ResponseEntity(
+                MessageDTO("You do not have permissions to access this resource"),
+                HttpStatus.UNAUTHORIZED
+            )
+            else ResponseEntity(teacherService.save(teacher), HttpStatus.OK)
         }
 
         @GetMapping
@@ -98,7 +109,10 @@ class TeacherController {
                     )]
                 )]
         )
-        fun getTeacher(@NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+        fun getTeacher(@CookieValue("jwt") jwt: String?, @NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+            if (jwt.isNullOrBlank()) {
+                return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+            }
             return ResponseEntity(teacherService.findById(id), HttpStatus.OK)
         }
 
@@ -142,8 +156,16 @@ class TeacherController {
                     )]
                 )]
         )
-        fun updateTeacher(@RequestBody teacher: Teacher): ResponseEntity<Any> {
-            return ResponseEntity(teacherService.update(teacher), HttpStatus.OK)
+        fun updateTeacher(@CookieValue("jwt") jwt: String?, @RequestBody teacher: Teacher): ResponseEntity<Any> {
+            if (jwt.isNullOrBlank()) {
+                return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+            }
+            val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+            return if (body["role"] == "STUDENT") ResponseEntity(
+                MessageDTO("You do not have permissions to access this resource"),
+                HttpStatus.UNAUTHORIZED
+            )
+            else ResponseEntity(teacherService.update(teacher), HttpStatus.OK)
         }
 
         @DeleteMapping
@@ -188,7 +210,15 @@ class TeacherController {
                     )]
                 )]
         )
-        fun deleteTeacher(@NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+        fun deleteTeacher(@CookieValue("jwt") jwt: String?, @NotBlank @RequestParam id: Long): ResponseEntity<Any> {
+            if (jwt.isNullOrBlank()) {
+                return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+            }
+            val body = Jwts.parser().setSigningKey("secret".encodeToByteArray()).parseClaimsJws(jwt).body
+            if (body["role"] == "STUDENT") return ResponseEntity(
+                MessageDTO("You do not have permissions to access this resource"),
+                HttpStatus.UNAUTHORIZED
+            )
             teacherService.deleteById(id)
             return ResponseEntity(MessageDTO("Teacher has been deleted successfully"), HttpStatus.OK)
         }
@@ -211,7 +241,10 @@ class TeacherController {
                     ]
                 )]
         )
-        fun getAll(): ResponseEntity<List<Teacher>> {
+        fun getAll(@CookieValue("jwt") jwt: String?): ResponseEntity<Any> {
+            if (jwt.isNullOrBlank()) {
+                return ResponseEntity(MessageDTO("It is not authenticated. Please log in"), HttpStatus.UNAUTHORIZED)
+            }
             return ResponseEntity(teacherService.readAll(), HttpStatus.OK)
         }
     }
